Adding several new functions to more expose more plugin container functionality
The inventory function takes an options map which in turn is used to filter
resources returned from the function. You can filter on availability, resource
category (i.e., platform, server, or service), resource type, and plugin.

There are a couple functions for dealing with operations.

There was previously a discovery function which has been replaced by discover.
If invoked with no args, it does both a server and service scan. You can also
specify a single arg to perform only one or the other.