inventory-filter now an arbitrary fn with the :fn option
In addition to the options that you can specify in the filter map that is
passed to the inventory fn, there is now a :fn option that expects a
function that receives a Resource object as an argument and returns true if the
resource should be included in the results.

Refactored the resource-category fn making it public and making it return a
keyword instead of an AvailabilityType object.

Lastly, adding more fn docstrings.